Bacheor thesis
*Title: Simulation of eddy currents in the Seville spherical tokamak.

In this work I studied numerically using the Fiesta code (MATLAB) the configuration of a future tokamak that will be built at the University of Seville (Spain). 4 different coilset and vessel configuration were studied, taking into account the current demands, the eddy currents and the mechanical stresses the vessel has to withstand. Also, the fundamentals of toroidal magnetic fusion devices are reviewed.