Master's thesis
*Title: Simulation of a plasma break-down for the SMART tokamak

In this work I studied numerically (MATLAB) the first phase of the start-up of the SMART tokamak
of the University of Seville, optimizing the current waveforms and the pressure of the pre-fill gas to ensure break-down
(ionization) of the gas without assistance. Supervisors: Dr. Manuel García Muñoz and Dr. Carlos Soria del Hoyo